requirements.txt
PyYAML==5.4.1
rich==9.13.0